requirements.txt
biopython==1.79
nglview==3.0.3
ipywidgets==7.6.5
